GSWB-83 Updated deploy.sh to match image names from harbor and updated dockerfile to use restored packages from the package  image.

+++ deploy.sh
@@ -4,9 +4,11 @@ set -eux
 DockerVersion=1.0.0
 
 # tarball csproj files, sln files, and NuGet.config
-find . \( -name "*.csproj" -o -name "*.sln" -o -name "NuGet.config" \) -print0 \
+find . \( -name "*.csproj" -o -name "*.sln" -o -name "NuGet.docker.config" \) -print0 \
     | tar -cvf projectfiles.tar --null -T -
 
-docker build . -t gswb.api-gateway:$DockerVersion -t gswb.api-gateway:latest
-
-rm projectfiles.tar
+docker build . -t registry.guildswarm.org/backend-images/api-gateway:$DockerVersion -t registry.guildswarm.org/backend-images/api-gateway:latest
+if [[ $0 == 1 || 0 ]]
+then
+    rm projectfiles.tar
+fi
